Changed Events to Doable, implement timeblock
